Implement plugins using extraction regex instead of js-code.
Most plugins are based on simple regular expression based extraction
of the strip url. This change adds the ability for a plugin to specify
an "extractRegex" and (optionally) "extractReplacement" to be used
instead of an extraction script.

This is the first step to support updating strip plugins without
the necessity to download and execute javascript code.